# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: PromoGo API
  version: "1.0"
  description: >
    PromoGo's canonical HTTP API contract: health/readiness, the store-scoped
    1C/POS webhook API (accrual, redemption, client lookup, balance), the
    mobile-customer OTP/session API, and the retailer-staff/platform-admin
    OIDC + RBAC API. Every operation here corresponds 1:1 to a registered
    route in internal/httpserver's route table (see
    internal/httpserver/routes.go) — this parity is enforced by
    TestRouteOpenAPIParity.
  license:
    name: Proprietary
    url: https://github.com/MirzaDgtu/PromoGo
servers:
  - url: /
    description: Relative to wherever the API is deployed; there is no fixed public host.
security:
  - {}
...
paths:
...
  /api/v1/transactions:
    post:
      operationId: createTransaction
      tags: [Store]
      summary: Accrue points for a completed sale (1C webhook)
      description: >
        Idempotent on (store, transaction_id): a retried request with the
        same transaction_id and parameters returns the original result
        (`replayed: true`) rather than crediting points twice. Requires the
        `transactions.write` scope.
      security:
        - storeApiKey: [transactions.write]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AccrueRequest" }
            examples:
              1c-webhook:
                summary: Typical 1C webhook payload
                value:
                  transaction_id: "1C-000123456"
                  phone: "+79261234567"
                  amount: "1250.00"
      responses:
        "200":
          description: Accrual processed (or replayed).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TransactionResponse" }
              examples:
                accrued:
                  value: { client_id: 42, points_earned: 125, balance: 640 }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /api/v1/transactions/redeem:
    post:
      operationId: redeemTransaction
      tags: [Store]
      summary: Redeem points at checkout (1C webhook)
      description: >
        Idempotent on (store, transaction_id), like accrual. Requires the
        `transactions.write` scope.
      security:
        - storeApiKey: [transactions.write]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RedeemRequest" }
      responses:
        "200":
          description: Redemption processed (or replayed).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RedeemResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: client_id does not belong to this store.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422":
          description: Insufficient balance to cover the requested redemption.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /api/v1/transactions/refund:
    post:
      operationId: refundTransaction
      tags: [Store]
      summary: Partially or fully refund a prior accrual/redeem (1C webhook)
      description: >
        Idempotent on (store, transaction_id), like accrual/redeem.
        original_transaction_id/original_transaction_type identify the
        transaction being reversed; original_transaction_type is only
        required when original_transaction_id was used for both an accrual
        and a redeem under this store. Multiple partial refunds against the
        same original are supported — see refunded_amount_total/
        fully_refunded on the response. Requires the `transactions.write`
        scope.
      security:
        - storeApiKey: [transactions.write]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefundRequest" }
            examples:
              partial-refund:
                summary: Partial refund of an accrual
                value:
                  transaction_id: "1C-000123456-R1"
                  original_transaction_id: "1C-000123456"
                  amount: "500.00"
      responses:
        "200":
          description: Refund processed (or replayed).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RefundResponse" }
        "400":
          description: Malformed request, or original_transaction_type required to disambiguate (code "ambiguous_original_transaction"), or the original transaction is itself a refund (code "cannot_refund_refund").
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: original_transaction_id does not exist under this store.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422":
          description: Refund amount exceeds the original's remaining refundable amount (code "over_refund"), or insufficient balance.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /api/v1/clients/resolve-qr:
    post:
      operationId: resolveQR
      tags: [Store]
      summary: Exchange a customer-issued QR payload for a store-scoped client (POS)
      description: >
        Atomically consumes a one-time QR payload minted by POST /me/qr,
        resolving (and creating/linking if this is the client's first visit
        to this store) a store-scoped client. Requires the `clients.lookup`
        scope. See DEC-011.
      security:
        - storeApiKey: [clients.lookup]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResolveQRRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResolveQRResponse" }
        "400":
          description: Malformed payload (code "qr_malformed").
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "410": { $ref: "#/components/responses/Gone" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /api/v1/clients/lookup:
    get:
      operationId: lookupClientByPhone
      tags: [Store]
      summary: Look up a client by phone (1C/POS)
      description: >
        Resolves a client_id and current balance for a phone already
        registered under the requesting store. Requires the
        `clients.lookup` scope.
      security:
        - storeApiKey: [clients.lookup]
      parameters:
        - name: phone
          in: query
          required: true
          schema: { type: string }
          example: "+79261234567"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClientBalance" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /api/v1/clients/{id}/balance:
    get:
      operationId: getClientBalance
      tags: [Store]
      summary: Get a client's point balance (1C/POS)
      description: 404s if id doesn't belong to the requesting store. Requires the `balances.read` scope.
      security:
        - storeApiKey: [balances.read]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClientBalance" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /api/v1/auth/otp/request:
    post:
      operationId: requestOTP
      tags: [Customer]
      summary: Request an OTP code for a phone number
      description: >
        Always 200 with the same body whether or not phone is already
        registered — cannot be used to enumerate accounts.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string, example: "+79261234567" }
      responses:
        "200":
          description: OTP sent (or would have been).
          content:
            application/json:
              schema:
                type: object
                properties: { status: { type: string, example: sent } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /api/v1/auth/otp/verify:
    post:
      operationId: verifyOTP
      tags: [Customer]
      summary: Verify an OTP code and obtain a session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, code]
              properties:
                phone: { type: string }
                code: { type: string, example: "123456" }
                consent_document_version:
                  type: string
                  description: Required on first registration; omit on a returning login.
                consent_source: { type: string, example: mobile_app }
            examples:
              mobile-verify:
                summary: Mobile app verifying a freshly-requested code
                value: { phone: "+79261234567", code: "123456" }
      responses:
        "200":
          description: Verified — access/refresh token pair issued.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: Invalid or expired code.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Account blocked.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /api/v1/auth/refresh:
    post:
      operationId: refreshCustomerSession
      tags: [Customer]
      summary: Rotate a refresh token
      description: Presenting an already-rotated refresh token is treated as compromise — every session for that account is revoked.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties: { refresh_token: { type: string } }
      responses:
        "200":
          description: New token pair issued.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: Invalid, expired, or revoked session.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/auth/logout:
    post:
      operationId: customerLogout
      tags: [Customer]
      summary: Revoke a single session
      description: Idempotent — an unknown/already-revoked token still returns 204.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties: { refresh_token: { type: string } }
      responses:
        "204": { description: Logged out. }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/auth/logout-all:
    post:
      operationId: customerLogoutAll
      tags: [Customer]
      summary: Revoke every session for the caller's account
      security:
        - customerAccessToken: []
      responses:
        "204": { description: All sessions revoked. }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/me:
    get:
      operationId: getMe
      tags: [Customer]
      summary: Current customer profile
      security:
        - customerAccessToken: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerProfile" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/me/balance:
    get:
      operationId: getMyBalance
      tags: [Customer]
      summary: Current customer's point balance, per linked store
      security:
        - customerAccessToken: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  balances:
                    type: array
                    items: { $ref: "#/components/schemas/MeBalanceItem" }
              examples:
                mobile-balance:
                  value: { balances: [{ store_id: 1, client_id: 42, balance: 640 }] }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /api/v1/me/transactions:
    get:
      operationId: getMyTransactions
      tags: [Customer]
      summary: Current customer's transaction history across all linked stores
      description: Keyset-paginated, newest first.
      security:
        - customerAccessToken: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
        - name: cursor
          in: query
          schema: { type: string }
          description: Opaque continuation token from a previous response's next_cursor.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MeTransactionsPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/me/qr:
    post:
      operationId: issueQR
      tags: [Customer]
      summary: Mint a new one-time QR payload for identification at POS
      description: >
        The payload carries no PII and is one-time-use (see DEC-011).
        Re-issuing before the previous QR's issue cooldown elapses returns
        429 with Retry-After.
      security:
        - customerAccessToken: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IssueQRResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/me/devices:
    post:
      operationId: registerDevice
      tags: [Customer]
      summary: Register or refresh an FCM push token for the caller's account
      security:
        - customerAccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterDeviceRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RegisterDeviceResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/me/devices/{deviceID}:
    delete:
      operationId: revokeDevice
      tags: [Customer]
      summary: Revoke one of the caller's own devices
      description: 404s (not 403) if deviceID belongs to a different account, so the response never confirms the ID exists.
      security:
        - customerAccessToken: []
      parameters:
        - name: deviceID
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "204": { description: Revoked }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/staff/auth/oidc:
    post:
      operationId: staffOIDCLogin
      tags: [Staff Auth]
      summary: Exchange a verified OIDC ID token for a PromoGo staff access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id_token]
              properties: { id_token: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { access_token: { type: string } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: Invalid ID token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Account disabled or no active organization membership.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /api/v1/admin/organizations:
    post:
      operationId: createOrganization
      tags: [Admin - Organizations]
      summary: Create an Organization
      description: Requires the global `organizations.manage` permission (any organization) — a bootstrap operation.
      security:
        - staffAccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties: { name: { type: string, maxLength: 255 } }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
...
components:
...
  responses:
    BadRequest:
      description: Malformed request body, path, or query parameters.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, invalid, or expired credential.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  securitySchemes:
    storeApiKey:
      type: http
      scheme: bearer
      description: >
        A per-store webhook API key (see Admin - API Keys), scoped to one
        or more of transactions.write / clients.lookup / balances.read. A
        store's legacy single-key credential (predating scopes) is treated
        as implicitly holding every scope.
    customerAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived customer access token from /auth/otp/verify or /auth/refresh.
    staffAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Short-lived staff access token from /staff/auth/oidc. Every
        operation under /admin also requires a specific RBAC permission —
        see domain.Permission and each operation's description — checked
        against the caller's active StaffMembership rows, resolved fresh on
        every request (never trusted from token claims).